LayoutTests/media/audio-playback-volume-changes-with-restrictions-and-user-gestures-expected.txt


RUN(internals.setMediaElementRestrictions(audio, "RequireUserGestureForAudioRateChange"))
RUN(audio.src = findMediaFile("audio", "content/test"))
RUN(audio.volume = 1)
RUN(audio.play().then(failTest).catch(didPreventPlayback))
Setting volume to 0 to allow playback
RUN(audio.volume = 0)
RUN(audio.play().then(didPlaySilentAudio).catch(failTest))
Ensure setting volume to non-zero value with a user gesture does not pause media in the presence of audio rate change restrictions
RUN(audio.volume = 0.2)
EXPECTED (audio.paused == 'false') OK
END OF TEST
